Moved adjustPdfHeightToScale to DistributionsChart/utils
